+++ backend/app/routers/health.py
@@ -1,4 +1,5 @@
 from fastapi import APIRouter
+import subprocess
 
 router = APIRouter()
 
@@ -8,4 +9,18 @@ def read_root():
 
 @router.get("/health")
 def health_check():
-    return {"status": "ok"}
+    def check_cmd(cmd):
+        try:
+            result = subprocess.run(cmd, capture_output=True, text=True, check=True)
+            return True, result.stdout.strip()
+        except Exception as e:
+            return False, str(e)
+
+    yt_dlp_ok, yt_dlp_msg = check_cmd(["yt-dlp", "--version"])
+    ffmpeg_ok, ffmpeg_msg = check_cmd(["ffmpeg", "-version"])
+
+    return {
+        "status": "ok" if yt_dlp_ok and ffmpeg_ok else "error",
+        "yt_dlp": {"ok": yt_dlp_ok, "msg": yt_dlp_msg},
+        "ffmpeg": {"ok": ffmpeg_ok, "msg": ffmpeg_msg},
+    }
