+++ backend/requirements.txt
@@ -2,3 +2,4 @@ fastapi==0.110.2
 uvicorn[standard]==0.29.0
 yt-dlp==2024.4.9 
 python-dotenv==1.0.1 
+ffmpeg-python
